send eth instead of wei to the backend improve testing clean up
