+++ .example.env
@@ -1,8 +1,6 @@
 MAIN=https://mainnet.infura.io/v3/
 TEST=https://rinkeby.infura.io/v3/
 
-ACCOUNT_PRIVATE_KEY=
-
 DB_HOST=
 DB_USER=
 DB_PASSWORD=
