+++ main.go
@@ -1,17 +1,14 @@
 package main
 
 import (
-	"encoding/json"
 	"ethereum-service/database"
-	"ethereum-service/internal"
 	"ethereum-service/internal/config"
+	"ethereum-service/internal/controller"
 	repository "ethereum-service/internal/repository"
-	"ethereum-service/model"
 	"ethereum-service/openApi"
 	"ethereum-service/services"
 	"fmt"
 	"log"
-	"math/big"
 	"net/http"
 	"strconv"
 
@@ -79,9 +76,9 @@ func checkAllAddresses() {
 	for _, s := range paymentIntents {
 		switch s.Mode {
 		case "main":
-			internal.CheckBalance(clientMain, &s)
+			controller.CheckBalance(clientMain, &s)
 		case "test":
-			internal.CheckBalance(clientTest, &s)
+			controller.CheckBalance(clientTest, &s)
 		default:
 			log.Fatal("Mode not supported!")
 		}
@@ -102,21 +99,6 @@ func InitializeRouter() *mux.Router {
 	sh := http.StripPrefix("/api/swaggerui/", http.FileServer(http.Dir("./swaggerui/")))
 	router.PathPrefix("/api/swaggerui/").Handler(sh)
 
-	router.HandleFunc("/test", test).Methods("GET", "OPTIONS")
 	//router.HandleFunc("/payment-intent", paymentIntent).Methods("POST", "OPTIONS")
 	return router
-}
-
-func test(w http.ResponseWriter, r *http.Request) {
-	fmt.Println(getTest())
-	err := json.NewEncoder(w).Encode(getTest())
-
-	var payment model.Payment
-	database.DB.Preload("CurrentPaymentState").First(&payment)
-	payment.UpdatePaymentState("paid", big.NewInt(2))
-	database.DB.Save(&payment)
-	if err != nil {
-		return
-	}
-	return
 }
