--- utils/utils_env.go
@@ -1,1 +1,0 @@
-package utils
