+++ utils/utils_eth.go
@@ -1,6 +1,12 @@
 package utils
 
-import "math/big"
+import (
+	"crypto/ecdsa"
+	"math/big"
+	"strings"
+
+	"github.com/ethereum/go-ethereum/crypto"
+)
 
 func GetETHFromWEI(amount *big.Int) *big.Float {
 	return big.NewFloat(0).Quo(new(big.Float).SetInt(amount), big.NewFloat(1000000000000000000))
@@ -19,4 +25,11 @@ func GetWEIFromETH(val *float64) *big.Int {
 		final.Add(final, big.NewInt(1))
 	}
 	return final
+}
+
+func GetPrivateKey(key string) (*ecdsa.PrivateKey, error) {
+	if strings.HasPrefix(key, "0x") {
+		key = key[2:]
+	}
+	return crypto.HexToECDSA(key)
 }
